+++ server/db/migrations/20180722005928_create_card_label_table.js
@@ -2,7 +2,7 @@ module.exports.up = (knex) =>
   knex.schema.createTable('card_label', (table) => {
     /* Columns */
 
-    table.bigInteger('id').primary();
+    table.increments('id').primary();
 
     table.bigInteger('card_id').notNullable();
     table.bigInteger('label_id').notNullable();
