+++ server/db/migrations/20180722006570_create_task_table.js
@@ -2,7 +2,7 @@ module.exports.up = (knex) =>
   knex.schema.createTable('task', (table) => {
     /* Columns */
 
-    table.bigInteger('id').primary();
+    table.increments('id').primary();
 
     table.bigInteger('card_id').notNullable();
