+++ server/db/migrations/20180722006688_create_attachment_table.js
@@ -2,7 +2,7 @@ module.exports.up = (knex) =>
   knex.schema.createTable('attachment', (table) => {
     /* Columns */
 
-    table.bigInteger('id').primary();
+    table.increments('id').primary();
 
     table.bigInteger('card_id').notNullable();
     table.bigInteger('creator_user_id').notNullable();
